requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
crispy-bootstrap5==2025.4
cryptography==44.0.3
defusedxml==0.7.1
Django==5.2.1
django-bootstrap5==25.1
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-nested-routers==0.94.2
flake8==7.2.0
humanize==4.12.3
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.10.1
pyTelegramBotAPI==4.27.0
pytest==8.3.5
python-dotenv==1.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
